+++ packages/helpers/src/index.ts
@@ -1,3 +1,5 @@
+export { AnyFn } from "./once";
+
 import once from "./once";
 import preferDefault from "./preferDefault";