+++ packages/helpers/types/once.d.ts
@@ -1,3 +1,2 @@
-declare type AnyFn = (...args: Array<any>) => Promise<any>;
+export declare type AnyFn = (...args: Array<any>) => Promise<any>;
 export default function once(fn: AnyFn): AnyFn;
-export {};