+++ packages/router/src/index.ts
@@ -2,5 +2,6 @@ export * from "./types";
 
 import curi from "./curi";
 import pathname from "./interactions/pathname";
+import once from "./deprecatedOnce";
 
-export { curi, pathname };
+export { curi, pathname, once };