+++ packages/router/types/index.d.ts
@@ -1,4 +1,5 @@
 export * from "./types";
 import curi from "./curi";
 import pathname from "./interactions/pathname";
-export { curi, pathname };
+import once from "./deprecatedOnce";
+export { curi, pathname, once };